requirements.txt
# Web Frameworks
fastapi==0.109.0
uvicorn[standard]==0.27.0
streamlit==1.31.0
python-multipart==0.0.6

# Azure SDKs
azure-cosmos==4.5.1
azure-identity==1.15.0
openai==1.12.0
azure-core==1.30.0

# LLM and ML
langchain==0.1.6
langchain-openai==0.0.5
langchain-community==0.0.19
sentence-transformers==2.3.1
tiktoken==0.5.2

# Vector Store and Embeddings
numpy==1.26.3
faiss-cpu==1.7.4

# Data Processing
pandas==2.2.0
pydantic==2.6.0
pydantic-settings==2.1.0

# MCP (Model Context Protocol)
mcp==0.9.0
httpx==0.26.0
websockets==12.0

# Utilities
python-dotenv==1.0.1
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Monitoring and Logging
prometheus-client==0.19.0
structlog==24.1.0

# Development
pytest==8.0.0
pytest-asyncio==0.23.4
black==24.1.1
ruff==0.1.15
